+++ NEWS.md
@@ -10,7 +10,9 @@
 - Relative bias and relative MAE as evaluation metrics
 - Drought-related leaf senescence only occurring if 'StemPLC' increases
 - Embolized sapwood proportion discounted from sapwood maintenance respiration 
-- Evaluation of DBH and Height series
+- Evaluation of diameter increment (DI), DBH and Height series
+- Output growth rates not relative to sapwood area
+
 
 # Version 2.6.0
 - Nitrogen content for leaves, sapwood and fine roots added. 'Nleaf' replaces 'Narea' as the latter can be calculated from 'Nleaf' using 'SLA'.